Reworked XMPP stream element classes (again)
It's necessary we are able to distinguish between Stanzas and
non-Stanzas (called PlainStremElement) here. Previously this was only
able with instanceof checks. With the new class hierarchy, the
information is available at compile time.

This allows us to later design the XMPPConnection API like this

XMPPConnection.send(Stanza)
XMPPConnection.send(PlainStreamElement)

Note that was is called Stanza in the above example is currently called
Packet in Smack.